+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/codewhisperer/util/JavaCodeWhispererFileCrawler.kt
@@ -7,7 +7,6 @@ import com.intellij.openapi.application.runReadAction
 import com.intellij.openapi.fileEditor.FileEditorManager
 import com.intellij.openapi.module.ModuleUtilCore
 import com.intellij.openapi.project.rootManager
-import com.intellij.openapi.roots.TestSourcesFilter
 import com.intellij.openapi.vfs.VfsUtil
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.psi.PsiClassOwner
@@ -24,22 +23,6 @@ object JavaCodeWhispererFileCrawler : CodeWhispererFileCrawler() {
     override val testFilenamePattern: Regex = """(?:Test([^/\\]+)\.java|([^/\\]+)Test\.java)$""".toRegex()
 
     override fun guessSourceFileName(tstFileName: String): String = tstFileName.substring(0, tstFileName.length - "Test.java".length) + ".java"
-
-    override fun listRelevantFilesInEditors(psiFile: PsiFile): List<VirtualFile> {
-        val targetFile = psiFile.virtualFile
-
-        val openedFiles = FileEditorManager.getInstance(psiFile.project).openFiles.toList().filter {
-            it.name != psiFile.virtualFile.name &&
-                it.extension == psiFile.virtualFile.extension &&
-                !TestSourcesFilter.isTestSources(it, psiFile.project)
-        }
-
-        val fileToFileDistanceList = openedFiles.map {
-            return@map it to CodeWhispererFileCrawler.getFileDistance(targetFile = targetFile, candidateFile = it)
-        }
-
-        return fileToFileDistanceList.sortedBy { it.second }.map { it.first }
-    }
 
     override suspend fun listFilesImported(psiFile: PsiFile): List<VirtualFile> {
         if (psiFile !is PsiJavaFile) return emptyList()
